refactor: extract selected RRT joint space
