+++ dimos/manipulation/planning/planners/rrt_planner.py
@@ -26,18 +26,15 @@ from typing import TYPE_CHECKING
 
 import numpy as np
 
-from dimos.manipulation.planning.groups.identifiers import (
-    is_global_joint_name,
-    local_joint_name_from_global,
-    make_global_joint_name,
-)
 from dimos.manipulation.planning.groups.models import PlanningGroupSelection
+from dimos.manipulation.planning.planners.selected_joint_space import (
+    SelectedJointSpace,
+    normalize_selection_target,
+)
 from dimos.manipulation.planning.spec.enums import PlanningStatus
 from dimos.manipulation.planning.spec.models import (
     JointPath,
-    LocalModelJointName,
     PlanningResult,
-    RobotName,
     WorldRobotID,
 )
 from dimos.manipulation.planning.spec.protocols import WorldSpec
@@ -68,16 +65,6 @@ class TreeNode:
             path.append(node.config)
             node = node.parent
         return list(reversed(path))
-
-
-@dataclass(frozen=True)
-class _SelectedRobotState:
-    robot_id: WorldRobotID
-    robot_name: RobotName
-    local_joint_names: list[LocalModelJointName]
-    base_positions_by_local_name: dict[LocalModelJointName, float]
-    lower_limits_by_local_name: dict[LocalModelJointName, float]
-    upper_limits_by_local_name: dict[LocalModelJointName, float]
 
 
 class RRTConnectPlanner:
@@ -201,18 +188,18 @@ class RRTConnectPlanner:
 
         selected_joint_names = list(selection.joint_names)
         try:
-            normalized_start = _normalize_selection_target(selection, start, "start")
+            normalized_start = normalize_selection_target(selection, start, "start")
         except ValueError as exc:
             return _create_failure_result(PlanningStatus.INVALID_START, str(exc))
         try:
-            normalized_goal = _normalize_selection_target(selection, goal, "goal")
+            normalized_goal = normalize_selection_target(selection, goal, "goal")
         except ValueError as exc:
             return _create_failure_result(PlanningStatus.INVALID_GOAL, str(exc))
         try:
-            robot_states = _build_selected_robot_states(world, selection)
+            selected_space = SelectedJointSpace.from_world(world, selection)
             q_start = np.asarray(normalized_start.position, dtype=np.float64)
             q_goal = np.asarray(normalized_goal.position, dtype=np.float64)
-            lower, upper = _selected_joint_limits(robot_states, selected_joint_names)
+            lower, upper = selected_space.joint_limits()
         except ValueError as exc:
             return _create_failure_result(PlanningStatus.NO_SOLUTION, str(exc))
 
@@ -227,21 +214,19 @@ class RRTConnectPlanner:
                 "Goal configuration is outside joint limits",
             )
 
-        if not _selected_config_collision_free(world, robot_states, selected_joint_names, q_start):
+        if not selected_space.config_collision_free(world, q_start):
             return _create_failure_result(
                 PlanningStatus.COLLISION_AT_START,
                 "Start configuration is in collision",
             )
-        if not _selected_config_collision_free(world, robot_states, selected_joint_names, q_goal):
+        if not selected_space.config_collision_free(world, q_goal):
             return _create_failure_result(
                 PlanningStatus.COLLISION_AT_GOAL,
                 "Goal configuration is in collision",
             )
 
-        if _selected_edge_collision_free(
+        if selected_space.edge_collision_free(
             world,
-            robot_states,
-            selected_joint_names,
             q_start,
             q_goal,
             self._collision_step_size,
@@ -265,18 +250,16 @@ class RRTConnectPlanner:
 
             sample = np.random.uniform(lower, upper)
             extended = self._extend_selected_tree(
+                selected_space,
                 world,
-                robot_states,
-                selected_joint_names,
                 start_tree,
                 sample,
                 self._step_size,
             )
             if extended is not None:
                 connected = self._connect_selected_tree(
+                    selected_space,
                     world,
-                    robot_states,
-                    selected_joint_names,
                     goal_tree,
                     extended.config,
                     self._connect_step_size,
@@ -285,10 +268,8 @@ class RRTConnectPlanner:
                     path = self._extract_path(extended, connected, selected_joint_names)
                     if trees_swapped:
                         path = list(reversed(path))
-                    path = _simplify_selected_path(
+                    path = selected_space.simplify_path(
                         world,
-                        robot_states,
-                        selected_joint_names,
                         path,
                         self._collision_step_size,
                     )
@@ -306,9 +287,8 @@ class RRTConnectPlanner:
 
     def _extend_selected_tree(
         self,
+        selected_space: SelectedJointSpace,
         world: WorldSpec,
-        robot_states: list[_SelectedRobotState],
-        selected_joint_names: list[str],
         tree: list[TreeNode],
         target: NDArray[np.float64],
         step_size: float,
@@ -322,10 +302,8 @@ class RRTConnectPlanner:
         else:
             new_config = nearest.config + step_size * (diff / dist)
 
-        if _selected_edge_collision_free(
+        if selected_space.edge_collision_free(
             world,
-            robot_states,
-            selected_joint_names,
             nearest.config,
             new_config,
             self._collision_step_size,
@@ -338,9 +316,8 @@ class RRTConnectPlanner:
 
     def _connect_selected_tree(
         self,
+        selected_space: SelectedJointSpace,
         world: WorldSpec,
-        robot_states: list[_SelectedRobotState],
-        selected_joint_names: list[str],
         tree: list[TreeNode],
         target: NDArray[np.float64],
         step_size: float,
@@ -348,9 +325,8 @@ class RRTConnectPlanner:
         """Try to connect a selected-joint tree to a target."""
         while True:
             result = self._extend_selected_tree(
+                selected_space,
                 world,
-                robot_states,
-                selected_joint_names,
                 tree,
                 target,
                 step_size,
@@ -522,265 +498,6 @@ class RRTConnectPlanner:
                 simplified = simplified[: i + 1] + simplified[j:]
 
         return simplified
-
-
-def _normalize_selection_target(
-    selection: PlanningGroupSelection,
-    target: JointState,
-    label: str,
-) -> JointState:
-    """Normalize a selected-joint target to global selection order."""
-    selected_global_names = list(selection.joint_names)
-    if not target.name:
-        if len(target.position) != len(selected_global_names):
-            raise ValueError(
-                f"{label} target has {len(target.position)} positions, "
-                f"expected {len(selected_global_names)}"
-            )
-        return JointState({"name": selected_global_names, "position": list(target.position)})
-
-    if len(target.name) != len(target.position):
-        raise ValueError(
-            f"{label} target has {len(target.name)} names but {len(target.position)} positions"
-        )
-
-    names = list(target.name)
-    global_flags = [is_global_joint_name(name) for name in names]
-    if any(global_flags) and not all(global_flags):
-        raise ValueError(f"{label} target mixes global and local joint names: {names}")
-
-    if all(global_flags):
-        expected_names = selected_global_names
-    else:
-        if len(selection.groups) != 1:
-            raise ValueError(
-                f"{label} target uses local joint names for a multi-group selection; "
-                "use global joint names"
-            )
-        expected_names = list(selection.groups[0].local_joint_names)
-
-    positions_by_name = dict(zip(names, target.position, strict=True))
-    missing = [name for name in expected_names if name not in positions_by_name]
-    if missing:
-        raise ValueError(f"{label} target is missing joints: {missing}")
-    extra = sorted(set(names) - set(expected_names))
-    if extra:
-        raise ValueError(f"{label} target has extra joints: {extra}")
-
-    ordered_positions = [float(positions_by_name[name]) for name in expected_names]
-    return JointState({"name": selected_global_names, "position": ordered_positions})
-
-
-def _build_selected_robot_states(
-    world: WorldSpec,
-    selection: PlanningGroupSelection,
-) -> list[_SelectedRobotState]:
-    robot_ids_by_name = _robot_ids_by_name(world, selection.robot_names)
-    selected_robot_states: list[_SelectedRobotState] = []
-    with world.scratch_context() as ctx:
-        for robot_name in selection.robot_names:
-            robot_id = robot_ids_by_name[robot_name]
-            config = world.get_robot_config(robot_id)
-            local_joint_names = list(config.joint_names)
-            current_state = world.get_joint_state(ctx, robot_id)
-            base_positions_by_local_name = _positions_by_local_name(
-                current_state,
-                robot_name,
-                local_joint_names,
-            )
-            lower, upper = world.get_joint_limits(robot_id)
-            if len(lower) != len(local_joint_names) or len(upper) != len(local_joint_names):
-                raise ValueError(
-                    f"Robot '{robot_name}' joint limits do not match configured joints"
-                )
-            selected_robot_states.append(
-                _SelectedRobotState(
-                    robot_id=robot_id,
-                    robot_name=robot_name,
-                    local_joint_names=local_joint_names,
-                    base_positions_by_local_name=base_positions_by_local_name,
-                    lower_limits_by_local_name=dict(
-                        zip(local_joint_names, lower.tolist(), strict=True)
-                    ),
-                    upper_limits_by_local_name=dict(
-                        zip(local_joint_names, upper.tolist(), strict=True)
-                    ),
-                )
-            )
-    return selected_robot_states
-
-
-def _positions_by_local_name(
-    joint_state: JointState,
-    robot_name: RobotName,
-    local_joint_names: list[LocalModelJointName],
-) -> dict[LocalModelJointName, float]:
-    if not joint_state.name:
-        if len(joint_state.position) != len(local_joint_names):
-            raise ValueError(
-                f"Current state for robot '{robot_name}' has {len(joint_state.position)} positions, "
-                f"expected {len(local_joint_names)}"
-            )
-        return dict(zip(local_joint_names, map(float, joint_state.position), strict=True))
-
-    positions_by_name = dict(zip(joint_state.name, joint_state.position, strict=True))
-    positions_by_local_name: dict[LocalModelJointName, float] = {}
-    for local_name in local_joint_names:
-        global_name = make_global_joint_name(robot_name, local_name)
-        if local_name in positions_by_name:
-            positions_by_local_name[local_name] = float(positions_by_name[local_name])
-        elif global_name in positions_by_name:
-            positions_by_local_name[local_name] = float(positions_by_name[global_name])
-        else:
-            raise ValueError(
-                f"Current state for robot '{robot_name}' is missing joint '{local_name}'"
-            )
-    return positions_by_local_name
-
-
-def _selected_joint_limits(
-    robot_states: list[_SelectedRobotState],
-    selected_joint_names: list[str],
-) -> tuple[NDArray[np.float64], NDArray[np.float64]]:
-    states_by_robot_name = {state.robot_name: state for state in robot_states}
-    lower: list[float] = []
-    upper: list[float] = []
-    for global_name in selected_joint_names:
-        robot_name, local_name = _split_selected_global_joint_name(
-            global_name, states_by_robot_name
-        )
-        state = states_by_robot_name[robot_name]
-        lower.append(state.lower_limits_by_local_name[local_name])
-        upper.append(state.upper_limits_by_local_name[local_name])
-    return np.asarray(lower, dtype=np.float64), np.asarray(upper, dtype=np.float64)
-
-
-def _selected_config_collision_free(
-    world: WorldSpec,
-    robot_states: list[_SelectedRobotState],
-    selected_joint_names: list[str],
-    selected_positions: NDArray[np.float64],
-) -> bool:
-    with world.scratch_context() as ctx:
-        projected_states = _project_selected_config(
-            robot_states, selected_joint_names, selected_positions
-        )
-        for robot_state in robot_states:
-            world.set_joint_state(ctx, robot_state.robot_id, projected_states[robot_state.robot_id])
-        return all(
-            world.is_collision_free(ctx, robot_state.robot_id) for robot_state in robot_states
-        )
-
-
-def _selected_edge_collision_free(
-    world: WorldSpec,
-    robot_states: list[_SelectedRobotState],
-    selected_joint_names: list[str],
-    start: NDArray[np.float64],
-    end: NDArray[np.float64],
-    step_size: float,
-) -> bool:
-    distance = float(np.linalg.norm(end - start))
-    steps = max(1, int(np.ceil(distance / step_size)))
-    for step in range(steps + 1):
-        ratio = step / steps
-        candidate = start + ratio * (end - start)
-        if not _selected_config_collision_free(
-            world, robot_states, selected_joint_names, candidate
-        ):
-            return False
-    return True
-
-
-def _project_selected_config(
-    robot_states: list[_SelectedRobotState],
-    selected_joint_names: list[str],
-    selected_positions: NDArray[np.float64],
-) -> dict[WorldRobotID, JointState]:
-    selected_positions_by_global_name = dict(
-        zip(selected_joint_names, selected_positions.tolist(), strict=True)
-    )
-    projected_states: dict[WorldRobotID, JointState] = {}
-    for robot_state in robot_states:
-        positions: list[float] = []
-        for local_name in robot_state.local_joint_names:
-            global_name = make_global_joint_name(robot_state.robot_name, local_name)
-            position = selected_positions_by_global_name.get(
-                global_name,
-                robot_state.base_positions_by_local_name[local_name],
-            )
-            positions.append(float(position))
-        projected_states[robot_state.robot_id] = JointState(
-            {"name": list(robot_state.local_joint_names), "position": positions}
-        )
-    return projected_states
-
-
-def _simplify_selected_path(
-    world: WorldSpec,
-    robot_states: list[_SelectedRobotState],
-    selected_joint_names: list[str],
-    path: JointPath,
-    collision_step_size: float,
-    max_iterations: int = 100,
-) -> JointPath:
-    if len(path) <= 2:
-        return path
-
-    simplified = list(path)
-    for _ in range(max_iterations):
-        if len(simplified) <= 2:
-            break
-        i = np.random.randint(0, len(simplified) - 2)
-        j = np.random.randint(i + 2, len(simplified))
-        start = np.asarray(simplified[i].position, dtype=np.float64)
-        end = np.asarray(simplified[j].position, dtype=np.float64)
-        if _selected_edge_collision_free(
-            world,
-            robot_states,
-            selected_joint_names,
-            start,
-            end,
-            collision_step_size,
-        ):
-            simplified = simplified[: i + 1] + simplified[j:]
-    return simplified
-
-
-def _robot_ids_by_name(
-    world: WorldSpec,
-    robot_names: tuple[RobotName, ...],
-) -> dict[RobotName, WorldRobotID]:
-    robot_ids_by_name: dict[RobotName, WorldRobotID] = {}
-    for robot_name in robot_names:
-        matches = [
-            robot_id
-            for robot_id in world.get_robot_ids()
-            if world.get_robot_config(robot_id).name == robot_name
-        ]
-        if not matches:
-            raise ValueError(f"Robot '{robot_name}' not found")
-        if len(matches) > 1:
-            raise ValueError(f"Robot name '{robot_name}' is not unique in planning world")
-        robot_ids_by_name[robot_name] = matches[0]
-    return robot_ids_by_name
-
-
-def _split_selected_global_joint_name(
-    global_name: str,
-    states_by_robot_name: dict[RobotName, _SelectedRobotState],
-) -> tuple[RobotName, LocalModelJointName]:
-    for robot_name, state in states_by_robot_name.items():
-        try:
-            local_name = local_joint_name_from_global(robot_name, global_name)
-        except ValueError:
-            continue
-        if local_name not in state.local_joint_names:
-            raise ValueError(
-                f"Selected joint '{global_name}' is not configured for robot '{robot_name}'"
-            )
-        return robot_name, local_name
-    raise ValueError(f"Selected joint '{global_name}' does not belong to a selected robot")
 
 
 # Result Helpers
